chore: refactor demo componetns
