+++ frontend/ui/src/components/tanstack-table/demo/inline-edit-demo.tsx
@@ -1,10 +1,10 @@
 "use client"
 
 import React, { useState, useEffect, useCallback } from "react"
-import { registerCellRenderer, cellRendererRegistry } from "./cell-renderers"
-import { DataTable } from "./components/DataTable"
+import { registerCellRenderer, cellRendererRegistry } from "../cell-renderers"
+import { DataTable } from "../components/DataTable"
 import { EyeIcon } from "lucide-react"
-import { RowAction } from "./types"
+import { RowAction } from "../types"
 
 // Custom rating cell renderer
 const RatingCell: React.FC<{ value: number }> = ({ value }) => {
@@ -134,7 +134,7 @@ const USER_TABLE_COLUMNS = [
     accessorKey: "email" as const,
     id: "email",
     enableSorting: true,
-    enableInlineEdit: true  
+    enableInlineEdit: true
   },
   {
     headingName: "Joined",
@@ -142,7 +142,7 @@ const USER_TABLE_COLUMNS = [
     accessorKey: "joinDate" as const,
     id: "joinDate",
     enableSorting: true,
-    enableInlineEdit: true, 
+    enableInlineEdit: true,
     // Date formatting
     format: {
       dateFormat: "YYYY-MM-DD HH:mm"
@@ -154,14 +154,14 @@ const USER_TABLE_COLUMNS = [
     accessorKey: "isActive" as const,
     id: "isActive",
     enableSorting: true,
-    enableInlineEdit: true 
+    enableInlineEdit: true
   },
   {
     headingName: "Tags",
     type: "Tag" as const,
     accessorKey: "tags" as const,
     id: "tags",
-    enableInlineEdit: true 
+    enableInlineEdit: true
   },
   {
     headingName: "Plan",
