+++ frontend/ui/src/components/tanstack-table/demo/tasks-demo.tsx
@@ -1,11 +1,9 @@
 "use client"
 
 import React, { useState, useEffect, useCallback } from "react"
-import { z } from 'zod'
-import { registerCellRenderer, cellRendererRegistry } from "./cell-renderers"
-import { RowAction } from "./types"
-import { DataTable } from "./components/DataTable"
-import { EyeIcon, PencilIcon } from "lucide-react"
+import { registerCellRenderer, cellRendererRegistry } from "../cell-renderers"
+import { RowAction } from "../types"
+import { DataTable } from "../components/DataTable"
 
 // Custom rating cell renderer (example of extending the table with a new column type)
 const RatingCell: React.FC<{ value: number }> = ({ value }) => {
@@ -119,95 +117,6 @@ const SAMPLE_USERS: User[] = [
     profileImage: "https://i.pravatar.cc/150?u=olivia"
   },
 ]
-
-
-// Define the form schema in the correct format for AutoForm
-const userEditSchema = {
-  formSchema: {
-    type: "object",
-    properties: {
-      name: {
-        type: "string",
-        minLength: 2,
-        title: "Name",
-      },
-      email: {
-        type: "string",
-        format: "email",
-        title: "Email Address",
-      },
-      isActive: {
-        type: "boolean",
-        title: "Account Status",
-      },
-      plan: {
-        type: "string",
-        enum: ["success", "pending", "failed"],
-        title: "Plan Status",
-      },
-      balance: {
-        type: "number",
-        minimum: 0,
-        title: "Account Balance",
-      },
-      rating: {
-        type: "string",
-        enum: ["1", "2", "3", "4", "5"],
-        title: "User Rating",
-      },
-    },
-    required: ["name", "email", "plan", "rating"],
-  },
-  fieldConfig: {
-    name: {
-      description: "User's full name",
-      inputProps: {
-        placeholder: "Enter user's full name",
-      },
-    },
-    email: {
-      description: "User's email address",
-      inputProps: {
-        placeholder: "user@example.com",
-      },
-    },
-    isActive: {
-      description: "Whether the user account is active or inactive",
-    },
-    plan: {
-      description: "Current subscription plan status",
-      fieldType: "select",
-      inputProps: {
-        options: [
-          { label: "Success", value: "success" },
-          { label: "Pending", value: "pending" },
-          { label: "Failed", value: "failed" },
-        ],
-      },
-    },
-    balance: {
-      description: "User's account balance in USD",
-      fieldType: "number",
-      inputProps: {
-        placeholder: "0.00",
-        prefix: "$",
-      },
-    },
-    rating: {
-      description: "User satisfaction rating from 1-5 stars",
-      fieldType: "select",
-      inputProps: {
-        options: [
-          { label: "★", value: "1" },
-          { label: "★★", value: "2" },
-          { label: "★★★", value: "3" },
-          { label: "★★★★", value: "4" },
-          { label: "★★★★★", value: "5" },
-        ],
-      },
-    },
-  },
-}
 
 // Column definitions
 const USER_TABLE_COLUMNS = [
@@ -333,19 +242,15 @@ const USER_TABLE_FACETS = [
 ]
 
 /**
- * Enhanced UsersTableDemo with Edit Capabilities
- * This demo extends the previous example by adding row editing functionality
+ * Enhanced UsersTableDemo - Example usage of the improved DataTable component
  */
-const EditableUsersTableDemo = () => {
+const EnhancedUsersTableDemo = () => {
   // Pagination state
   const [currentPage, setCurrentPage] = useState(0)
   const [pageSize, setPageSize] = useState(10)
   const [totalItems, setTotalItems] = useState(0)
 
-  // Store all users data (not just the current page)
-  const [allUsers, setAllUsers] = useState<User[]>([...SAMPLE_USERS])
-
-  // Data and loading state for current page
+  // Data and loading state
   const [users, setUsers] = useState<User[]>([])
   const [loading, setLoading] = useState(true)
 
@@ -361,12 +266,12 @@ const EditableUsersTableDemo = () => {
       await new Promise(resolve => setTimeout(resolve, 800))
 
       // Set the total number of items (for pagination calculations)
-      setTotalItems(allUsers.length)
+      setTotalItems(SAMPLE_USERS.length)
 
       // Get the slice of data for the current page
       const start = page * size
       const end = start + size
-      const paginatedUsers = allUsers.slice(start, end)
+      const paginatedUsers = SAMPLE_USERS.slice(start, end)
 
       setUsers(paginatedUsers)
     } catch (error) {
@@ -374,7 +279,7 @@ const EditableUsersTableDemo = () => {
     } finally {
       setLoading(false)
     }
-  }, [allUsers])
+  }, [])
 
   // Initial data load
   useEffect(() => {
@@ -398,19 +303,27 @@ const EditableUsersTableDemo = () => {
   }, [])
 
   /**
-   * Define row actions
+   * Define row actions - memoized to prevent recreating on each render
    */
-  const getRowActions = useCallback((row: User): RowAction[] => {
-    return [
-      {
-        label: "View Profile",
-        icon: <EyeIcon className="h-4 w-4" />,
-        onClick: () => {
-          console.log("Viewing user profile:", row.name)
-        }
-      }
-    ]
-  }, [])
+  const getRowActions = useCallback((row: User): RowAction[] => [
+    {
+      label: "View Profile",
+      onClick: () => alert(`View profile of ${row.name}`),
+      icon: <span className="mr-2">👤</span>
+    },
+    {
+      label: "Edit User",
+      onClick: () => alert(`Edit user ${row.name}`),
+      icon: <span className="mr-2">✏️</span>
+    },
+    {
+      label: "Delete User",
+      onClick: () => alert(`Delete user ${row.name}`),
+      icon: <span className="mr-2">🗑️</span>,
+      color: "text-red-500",
+      disabled: row.isActive // Disable delete for active users
+    },
+  ], [])
 
   // User detail component for expandable rows
   const renderUserDetails = useCallback((user: User) => (
@@ -445,61 +358,16 @@ const EditableUsersTableDemo = () => {
         </div>
       </div>
     </div>
-  ), [])
+  ), []);
 
   // Handle selection change
   const handleSelectionChange = useCallback((selectedRows: User[]) => {
-    console.log("Selected rows:", selectedRows)
-  }, [])
-
-  /**
-   * Handle row edit - this is the key new functionality
-   */
-  const handleRowEdit = useCallback((oldData: User, newData: any) => {
-    // Log the edit
-    console.log("Editing user:", oldData)
-    console.log("New data:", newData)
-
-    // Convert string values back to their original types if needed
-    const processedData = {
-      ...newData,
-      // Convert rating from string to number
-      rating: typeof newData.rating === 'string' ? parseInt(newData.rating, 10) : newData.rating
-    };
-
-    // Create updated user (preserving fields not in the form)
-    const updatedUser = {
-      ...oldData,
-      ...processedData,
-      // Ensure we preserve fields not in the form schema
-      id: oldData.id,
-      tags: oldData.tags,
-      profileImage: oldData.profileImage,
-      joinDate: oldData.joinDate
-    }
-
-    // Update the user in the full dataset
-    setAllUsers(prevUsers =>
-      prevUsers.map(user =>
-        user.id === updatedUser.id ? updatedUser : user
-      )
-    )
-
-    // Also update the current page if the edited user is in the current view
-    setUsers(prevPageUsers =>
-      prevPageUsers.map(user =>
-        user.id === updatedUser.id ? updatedUser : user
-      )
-    )
-
-    // Show success message
-    alert(`User ${newData.name} updated successfully!`)
-  }, [])
+    console.log("Selected rows:", selectedRows);
+  }, []);
 
   return (
     <div className="container mx-auto py-10">
-      <h1 className="text-2xl font-bold mb-4">Editable Users Table</h1>
-      <p className="text-gray-600 mb-6">This demo showcases the edit functionality with form validation</p>
+      <h1 className="text-2xl font-bold mb-4">Enhanced Users Management</h1>
 
       <DataTable
         columns={USER_TABLE_COLUMNS}
@@ -530,26 +398,17 @@ const EditableUsersTableDemo = () => {
         }}
 
         // Expandable rows
-        // expandableRows={{
-        //   render: renderUserDetails,
-        //   expandOnClick: false, // Disable automatic expansion on row click
-        //   singleExpand: true
-        // }}
-
-        // Edit functionality - new additions
-        enableRowEdit={true}
-        editFormSchema={userEditSchema.formSchema}
-        editFieldConfig={userEditSchema.fieldConfig}
-        onRowEdit={handleRowEdit}
-        editDialogTitle="Edit User Profile"
+        expandableRows={{
+          render: renderUserDetails,
+          expandOnClick: true,
+          singleExpand: true
+        }}
 
         // Events
         onSelectionChange={handleSelectionChange}
       />
-
-
     </div>
   )
 }
 
-export default EditableUsersTableDemo
+export default EnhancedUsersTableDemo
