+++ frontend/ui/src/components/tanstack-table/index.ts
@@ -1,9 +1,9 @@
 // src/components/tanstack-table/index.ts
-export { default as UsersTableDemo } from "./users-demo";
-export { default as TasksTableDemo } from "./tasks-demo";
-export { default as SidebarFilterDemo } from "./sidebarfilter-demo";
-export { default as EditableUsersTableDemo } from "./users-edit-demo";
-export { default as InlineEditDemo } from "./inline-edit-demo";
+export { default as UsersTableDemo } from "./demo/users-demo";
+export { default as TasksTableDemo } from "./demo/tasks-demo";
+export { default as SidebarFilterDemo } from "./demo/sidebarfilter-demo";
+export { default as EditableUsersTableDemo } from "./demo/users-edit-demo";
+export { default as InlineEditDemo } from "./demo/inline-edit-demo";
 
 // Export the main DataTable component
 export { DataTable } from "./components/DataTable";
